server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
# Pillow is optional - used for image compression
# If installation fails, the app will work without image compression
# pip install Pillow --only-binary :all:
